requirements.txt
django==5.2
djangorestframework==3.16.0
django-filter==25.1
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.10
pillow==11.2.1
pytest==8.3.5
pytest-django==4.11.1 